contelizer_task/requirements_dev.txt
black
pytest
pytest-django
ruff
